feractor: update styles and folder store
