+++ src/components/App/App.scss
@@ -1,5 +1,11 @@
 button {
   cursor: pointer;
+  outline: none;
+  border: none;
+}
+
+button:focus {
+  box-shadow: none !important;
 }
 
 button::-moz-focus-inner {
@@ -26,6 +32,7 @@ body {
   padding: 0;
   height: 100vh;
   width: 100vw;
+  background-color: #e9ecef;
 }
 
 #root {
@@ -37,6 +44,19 @@ body {
   padding: 0;
   // width: 100%;
   height: 100%;
+}
+
+button {
+  &:hover {
+    transform: scale(1.1);
+    transition: all 0.1s ease-in-out;
+    border-radius: 5px;
+    cursor: pointer;
+  }
+
+  &:active {
+    transform: scale(0.9);
+  }
 }
 
 // .wrapper {
