+++ src/components/Autoplay/Autoplay.jsx
@@ -1,10 +1,10 @@
 import { connect } from 'react-redux';
-import { finishGame } from '../../store/mainReducer';
+import { finishGame } from '../../store/main/mainReducer';
 import style from './Autoplay.module.scss';
 import PropTypes from "prop-types";
 import { getCountPairs } from '../../selectors/mainSelectors';
-import { playGame } from '../../store/autoplay/reducer';
-import { toggleIsMenuActive } from '../../store/burgerReducer';
+import { playGame } from '../../store/autoplay/middleware';
+import { toggleIsMenuActive } from '../../store/burger/burgerReducer';
 
 const Autoplay = ({ toggleIsMenuActive, finishGame, countPairs, playGame }) => {
