+++ src/components/Card/Card.scss
@@ -1,9 +1,5 @@
 .flip-container{
-   // perspective: 1000px;
-   // display: inline-block;
-   border: 1px solid white;
    border-radius: 5px;
-   // background: rgb(182, 143, 143);
    margin: 5px;
 
    &:hover{
