+++ src/components/Header/HeaderContainer.jsx
@@ -1,7 +1,7 @@
 import { useHotkeys } from "react-hotkeys-hook";
 import { connect } from 'react-redux';
 import { getIsActive, getIsMusicActive, getIsSoundActive } from '../../selectors/burgerSelectors';
-import { setSoundVolume, toggleIsMenuActive, toggleIsMusic, toggleIsSound } from '../../store/burgerReducer';
+import { setSoundVolume, toggleIsMenuActive, toggleIsMusic, toggleIsSound } from '../../store/burger/burgerReducer';
 import PropTypes from "prop-types";
 import Header from './Header';
 import Menu from '../Menu/Menu';
