+++ src/components/StatsPage/StatsPage.module.scss
@@ -29,6 +29,7 @@
 .header {
   font-weight: bold;
   outline: 1px solid black;
+  background-color: rgb(190, 190, 190);
 }
 
 .row:hover {
