+++ src/store/autoplay copy/reducer.js
@@ -1,33 +1,8 @@
-const BURGER_TOGGLE_IS_MENU_ACTIVE = 'BURGER_TOGGLE_IS_MENU_ACTIVE';
-const BURGER_TOGGLE_SOUND = 'BURGER_TOGGLE_SOUND';
-const BURGER_TOGGLE_MUSIC = 'BURGER_TOGGLE_MUSIC';
-const BURGER_SET_SOUND_VOLUME = 'BURGER_SET_SOUND_VOLUME';
-const BURGER_SET_CARDS_BG = 'BURGER_SET_CARDS_BG';
-const BURGER_SET_OPACITY_BG = 'BURGER_SET_OPACITY_BG';
+import { getInitialState } from "../utils/utils";
+import { BURGER_SET_CARDS_BG, BURGER_SET_OPACITY_BG, BURGER_SET_SOUND_VOLUME, BURGER_TOGGLE_IS_MENU_ACTIVE, BURGER_TOGGLE_MUSIC, BURGER_TOGGLE_SOUND } from "./actionTypes";
+import { initialState } from "./state";
 
-const initialState = {
-  isMenuActive: false,
-  isSoundActive: true,
-  isMusicActive: false,
-  volume: 0.5,
-  cardsBG: '255,255,255',
-  opacity: 0,
-};
-
-const getInitialState = () => {
-  let state;
-  const localState = JSON.parse(localStorage.getItem('burger'));
-  if (localState) {
-    state = localState;
-  } else {
-    localStorage.setItem('burger', JSON.stringify(initialState));
-    state = initialState;
-  }
-
-  return state;
-};
-
-const burgerReducer = (state = getInitialState(), action) => {
+const burgerReducer = (state = getInitialState('burger', initialState), action) => {
   let newState;
   switch (action.type) {
     case BURGER_TOGGLE_IS_MENU_ACTIVE:
@@ -83,17 +58,5 @@ const burgerReducer = (state = getInitialState(), action) => {
       return state;
   }
 };
-
-export const toggleIsMenuActive = () => ({ type: BURGER_TOGGLE_IS_MENU_ACTIVE });
-
-export const toggleIsSound = () => ({ type: BURGER_TOGGLE_SOUND });
-
-export const toggleIsMusic = () => ({ type: BURGER_TOGGLE_MUSIC });
-
-export const setSoundVolume = (volume) => ({ type: BURGER_SET_SOUND_VOLUME, volume });
-
-export const setCardsBG = (color) => ({ type: BURGER_SET_CARDS_BG, color });
-
-export const setOpacityBG = (opacity) => ({ type: BURGER_SET_OPACITY_BG, opacity });
 
 export default burgerReducer;
